Refactor: Move vector search tools
